testing/web-platform/tests/atomics/window-and-any-worker/wake/wake-zero/wake-zero.html

<!doctype html>
<meta charset=utf-8>
<title></title>

<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>

<body>
    <iframe id="myiframe" style="display: none" src="../empty.html" ></iframe>
</body>

<script>

// Copyright (C) 2017 Mozilla Corporation.  All rights reserved.
// This code is governed by the BSD license found in the LICENSE file.

/*---
description: >
  Test that Atomics.wake wakes zero waiters if that's what the count is.
---*/

window.addEventListener("load", function() {
  async_test(function(t) {
    let worker = new Worker('wake-zero.js');
    worker.onmessage = this.step_func_done(function(e) {
      let ret = e.data[0];
      assert_equals(ret, "timed-out");
    });
    let ia = new Int32Array(new SharedArrayBuffer(Int32Array.BYTES_PER_ELEMENT));
    worker.postMessage([ia]);
    setTimeout(() => assert_equals(0, Atomics.wake(ia, 0, 0)), 500); // Wake zero
  }, "atomics-dedicated-worker-wake-two");
  async_test(function(t) {
    let worker = new Worker('wake-zero-worker.js');
    worker.onmessage = this.step_func_done(function(e) {
      let ret = e.data[0];
      assert_equals(ret, "timed-out");
    });
    worker.postMessage([]);
  }, "atomics-dedicated-worker-wake-zero-from-worker");
  async_test(function(t) {
    let worker = new SharedWorker('wake-zero-shared-worker.js');
    worker.port.onmessage = this.step_func_done(function(e) {
      let ret = e.data[0];
      assert_equals(ret, "timed-out");
    });
    worker.port.start();
    worker.port.postMessage([]);
  }, "atomics-dedicated-worker-wake-zero-from-shared-worker");
  async_test(function(t) {
    let iframe = document.getElementById('myiframe');
    let contentWindow = iframe.contentWindow;
    let test = this;
    contentWindow.onmessage = function(e) {
      let worker = new Worker('wake-zero.js');
      worker.onmessage = test.step_func_done(function(e) {
        let ret = e.data[0];
        assert_equals(ret, "timed-out");
      });
      let ia = new Int32Array(new SharedArrayBuffer(Int32Array.BYTES_PER_ELEMENT));
      worker.postMessage([ia]);
      setTimeout(() => assert_equals(0, Atomics.wake(ia, 0, 0)), 500); // Wake zero
    }
    contentWindow.postMessage([], '*');
  }, "atomics-dedicated-worker-wake-zero-from-iframe");
});
</script>
